Add muxing models
Small refactor to be able to have typing and make the code
more readable
