+++ src/codegate/api/v1_models.py
@@ -267,26 +267,3 @@ class ModelByProvider(pydantic.BaseModel):
 
     def __str__(self):
         return f"{self.provider_name} / {self.name}"
-
-
-class MuxMatcherType(str, Enum):
-    """
-    Represents the different types of matchers we support.
-    """
-
-    # Always match this prompt
-    catch_all = "catch_all"
-
-
-class MuxRule(pydantic.BaseModel):
-    """
-    Represents a mux rule for a provider.
-    """
-
-    provider_id: str
-    model: str
-    # The type of matcher to use
-    matcher_type: MuxMatcherType
-    # The actual matcher to use. Note that
-    # this depends on the matcher type.
-    matcher: Optional[str] = None
